Make project creation unique in migration wizard
