+++ packages/ballerina-visualizer/src/views/BI/ProjectForm/embedded/integrator-form/components/AdvancedConfigurationSection.tsx
@@ -66,6 +66,12 @@ export interface AdvancedConfigurationSectionProps {
     organizations?: Organization[];
     /** Whether the section contains validation errors */
     hasError?: boolean;
+    /**
+     * Hides the Package Name field. For a destination that receives SEVERAL packages at
+     * once (the migration wizard's multi-project import) there is no single package to
+     * name — only the shared organization and version are the user's to pick.
+     */
+    hidePackageName?: boolean;
 }
 
 // ── Helpers ────────────────────────────────────────────────────────────────────
@@ -314,7 +320,8 @@ export function AdvancedConfigurationSection({
     packageNameError,
     projectHandleError,
     organizations,
-    hasError
+    hasError,
+    hidePackageName
 }: AdvancedConfigurationSectionProps) {
     const { isSigningIn, handleSignIn, handleCancelSignIn } = useSignIn();
 
@@ -357,19 +364,23 @@ export function AdvancedConfigurationSection({
             )}
             <SubSectionLabel>Ballerina Package</SubSectionLabel>
             <Note style={{ marginBottom: "16px" }}>
-                {createWithinProject
-                    ? `This ${isLibrary ? "library" : "integration"} is generated as a Ballerina package. Specify the package name and version to be assigned.`
-                    : `This ${isLibrary ? "library" : "integration"} is generated as a Ballerina package. Specify the organization, package name and version to be assigned.`}
+                {hidePackageName
+                    ? `Each ${isLibrary ? "library" : "integration"} is generated as a Ballerina package. Specify the organization and version to be assigned.`
+                    : createWithinProject
+                        ? `This ${isLibrary ? "library" : "integration"} is generated as a Ballerina package. Specify the package name and version to be assigned.`
+                        : `This ${isLibrary ? "library" : "integration"} is generated as a Ballerina package. Specify the organization, package name and version to be assigned.`}
             </Note>
-            <FieldGroup>
-                <TextField
-                    onTextChange={(value) => onChange({ packageName: sanitizePackageName(value) })}
-                    value={data.packageName}
-                    label="Package Name"
-                    errorMsg={packageNameError || undefined}
-                />
-                <Description>Specify the package name.</Description>
-            </FieldGroup>
+            {!hidePackageName && (
+                <FieldGroup>
+                    <TextField
+                        onTextChange={(value) => onChange({ packageName: sanitizePackageName(value) })}
+                        value={data.packageName}
+                        label="Package Name"
+                        errorMsg={packageNameError || undefined}
+                    />
+                    <Description>Specify the package name.</Description>
+                </FieldGroup>
+            )}
             {!createWithinProject && (
                 <FieldGroup>
                     <OrgField
