+++ packages/ballerina-visualizer/src/views/BI/ProjectForm/embedded/integrator-form/hooks/useSignIn.ts
@@ -54,7 +54,19 @@ export function useSignIn() {
     // Clean up on unmount.
     useEffect(() => () => clearSignInTimeout(), []);
 
+    /**
+     * The form this backs is also rendered outside the embedded host (the migration
+     * wizard), where the context default leaves `wsClient` an empty object. Calling
+     * through it would throw on click, so sign-in is simply unavailable there rather
+     * than broken.
+     */
+    const canRunCommand = typeof wsClient?.runCommand === "function";
+
     const handleSignIn = () => {
+        if (!canRunCommand) {
+            console.warn("Sign-in is unavailable: no command transport in this host.");
+            return;
+        }
         setIsSigningIn(true);
         clearSignInTimeout();
         timeoutRef.current = setTimeout(() => {
@@ -69,6 +81,9 @@ export function useSignIn() {
     const handleCancelSignIn = () => {
         setIsSigningIn(false);
         clearSignInTimeout();
+        if (!canRunCommand) {
+            return;
+        }
         wsClient
             .runCommand({ command: WICommandIds.CancelSignIn, args: [] })
             .catch((error) => console.error("Cancel sign-in command failed:", error));
