+++ packages/ballerina-visualizer/src/views/BI/wsManager/WsClient.ts
@@ -171,6 +171,12 @@ export class BiWsClient {
         return this.request("getProjectComponentNames", params);
     }
 
+    public getExistingProjectInfo(
+        params: { projectPath: string }
+    ): Promise<{ isProject?: boolean; name?: string } | null> {
+        return this.request("getExistingProjectInfo", params);
+    }
+
     public selectFileOrDirPath(params: any): Promise<any> {
         return this.request("selectFileOrDirPath", params);
     }
